Fix saveAndContinue package moving
